+++ sdk/resourcemanager/voiceservices/armvoiceservices/communicationsgateways_client.go
@@ -47,7 +47,7 @@ func NewCommunicationsGatewaysClient(subscriptionID string, credential azcore.To
 // BeginCreateOrUpdate - Create a CommunicationsGateway
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2023-01-31
+// Generated from API version 2023-04-03
 //   - resourceGroupName - The name of the resource group. The name is case insensitive.
 //   - communicationsGatewayName - Unique identifier for this deployment
 //   - resource - Resource create parameters.
@@ -70,7 +70,7 @@ func (client *CommunicationsGatewaysClient) BeginCreateOrUpdate(ctx context.Cont
 // CreateOrUpdate - Create a CommunicationsGateway
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2023-01-31
+// Generated from API version 2023-04-03
 func (client *CommunicationsGatewaysClient) createOrUpdate(ctx context.Context, resourceGroupName string, communicationsGatewayName string, resource CommunicationsGateway, options *CommunicationsGatewaysClientBeginCreateOrUpdateOptions) (*http.Response, error) {
 	req, err := client.createOrUpdateCreateRequest(ctx, resourceGroupName, communicationsGatewayName, resource, options)
 	if err != nil {
@@ -106,7 +106,7 @@ func (client *CommunicationsGatewaysClient) createOrUpdateCreateRequest(ctx cont
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2023-01-31")
+	reqQP.Set("api-version", "2023-04-03")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Accept"] = []string{"application/json"}
 	return req, runtime.MarshalAsJSON(req, resource)
@@ -115,7 +115,7 @@ func (client *CommunicationsGatewaysClient) createOrUpdateCreateRequest(ctx cont
 // BeginDelete - Delete a CommunicationsGateway
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2023-01-31
+// Generated from API version 2023-04-03
 //   - resourceGroupName - The name of the resource group. The name is case insensitive.
 //   - communicationsGatewayName - Unique identifier for this deployment
 //   - options - CommunicationsGatewaysClientBeginDeleteOptions contains the optional parameters for the CommunicationsGatewaysClient.BeginDelete
@@ -137,7 +137,7 @@ func (client *CommunicationsGatewaysClient) BeginDelete(ctx context.Context, res
 // Delete - Delete a CommunicationsGateway
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2023-01-31
+// Generated from API version 2023-04-03
 func (client *CommunicationsGatewaysClient) deleteOperation(ctx context.Context, resourceGroupName string, communicationsGatewayName string, options *CommunicationsGatewaysClientBeginDeleteOptions) (*http.Response, error) {
 	req, err := client.deleteCreateRequest(ctx, resourceGroupName, communicationsGatewayName, options)
 	if err != nil {
@@ -173,7 +173,7 @@ func (client *CommunicationsGatewaysClient) deleteCreateRequest(ctx context.Cont
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2023-01-31")
+	reqQP.Set("api-version", "2023-04-03")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Accept"] = []string{"application/json"}
 	return req, nil
@@ -182,7 +182,7 @@ func (client *CommunicationsGatewaysClient) deleteCreateRequest(ctx context.Cont
 // Get - Get a CommunicationsGateway
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2023-01-31
+// Generated from API version 2023-04-03
 //   - resourceGroupName - The name of the resource group. The name is case insensitive.
 //   - communicationsGatewayName - Unique identifier for this deployment
 //   - options - CommunicationsGatewaysClientGetOptions contains the optional parameters for the CommunicationsGatewaysClient.Get
@@ -222,7 +222,7 @@ func (client *CommunicationsGatewaysClient) getCreateRequest(ctx context.Context
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2023-01-31")
+	reqQP.Set("api-version", "2023-04-03")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Accept"] = []string{"application/json"}
 	return req, nil
@@ -239,7 +239,7 @@ func (client *CommunicationsGatewaysClient) getHandleResponse(resp *http.Respons
 
 // NewListByResourceGroupPager - List CommunicationsGateway resources by resource group
 //
-// Generated from API version 2023-01-31
+// Generated from API version 2023-04-03
 //   - resourceGroupName - The name of the resource group. The name is case insensitive.
 //   - options - CommunicationsGatewaysClientListByResourceGroupOptions contains the optional parameters for the CommunicationsGatewaysClient.NewListByResourceGroupPager
 //     method.
@@ -287,7 +287,7 @@ func (client *CommunicationsGatewaysClient) listByResourceGroupCreateRequest(ctx
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2023-01-31")
+	reqQP.Set("api-version", "2023-04-03")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Accept"] = []string{"application/json"}
 	return req, nil
@@ -304,7 +304,7 @@ func (client *CommunicationsGatewaysClient) listByResourceGroupHandleResponse(re
 
 // NewListBySubscriptionPager - List CommunicationsGateway resources by subscription ID
 //
-// Generated from API version 2023-01-31
+// Generated from API version 2023-04-03
 //   - options - CommunicationsGatewaysClientListBySubscriptionOptions contains the optional parameters for the CommunicationsGatewaysClient.NewListBySubscriptionPager
 //     method.
 func (client *CommunicationsGatewaysClient) NewListBySubscriptionPager(options *CommunicationsGatewaysClientListBySubscriptionOptions) *runtime.Pager[CommunicationsGatewaysClientListBySubscriptionResponse] {
@@ -347,7 +347,7 @@ func (client *CommunicationsGatewaysClient) listBySubscriptionCreateRequest(ctx 
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2023-01-31")
+	reqQP.Set("api-version", "2023-04-03")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Accept"] = []string{"application/json"}
 	return req, nil
@@ -365,7 +365,7 @@ func (client *CommunicationsGatewaysClient) listBySubscriptionHandleResponse(res
 // Update - Update a CommunicationsGateway
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2023-01-31
+// Generated from API version 2023-04-03
 //   - resourceGroupName - The name of the resource group. The name is case insensitive.
 //   - communicationsGatewayName - Unique identifier for this deployment
 //   - properties - The resource properties to be updated.
@@ -406,7 +406,7 @@ func (client *CommunicationsGatewaysClient) updateCreateRequest(ctx context.Cont
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2023-01-31")
+	reqQP.Set("api-version", "2023-04-03")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Accept"] = []string{"application/json"}
 	return req, runtime.MarshalAsJSON(req, properties)